ft-new test cases commit
