+++ package.json
@@ -5,7 +5,7 @@
   "main": "app.js",
   "scripts": {
     "start": "set DEBUG=app,app:* & nodemon app.js",
-    "test": "mocha --file api/tests/test.test.js --exclude '**/*.js  --timeout 5000"
+    "test": "mocha --file api/tests/test.js --exclude '**/*.js  --timeout 5000"
   },
   "repository": {
     "type": "git",
